Move remainder of core tests
