+++ mantidimaging/core/aggregate/aggregate.py
