+++ mantidimaging/core/aggregate/test/aggregate_test.py
@@ -11,9 +11,9 @@ from mantidimaging.core.aggregate import aggregate
 from mantidimaging.core.configs.recon_config import ReconstructionConfig
 from mantidimaging.core.io import loader
 from mantidimaging.core.io.saver import Saver
-from mantidimaging.tests.file_outputting_test_case import (
+from mantidimaging.test.file_outputting_test_case import (
         FileOutputtingTestCase)
-from mantidimaging.tests import test_helper as th
+from mantidimaging.test import test_helper as th
 
 
 class AggregateTest(FileOutputtingTestCase):
