+++ mantidimaging/core/imopr/test/utility_test.py
