+++ mantidimaging/core/parallel/test/exclusive_mem_test.py
@@ -5,7 +5,7 @@ import unittest
 import numpy.testing as npt
 
 from mantidimaging.core.utility.memory_usage import get_memory_usage_linux
-from mantidimaging.tests import test_helper as th
+from mantidimaging.test import test_helper as th
 
 from mantidimaging.core.parallel import exclusive_mem as esm
