+++ mantidimaging/core/parallel/test/two_shared_mem_test.py
@@ -5,7 +5,7 @@ import unittest
 import numpy.testing as npt
 
 from mantidimaging.core.utility.memory_usage import get_memory_usage_linux
-from mantidimaging.tests import test_helper as th
+from mantidimaging.test import test_helper as th
 
 from mantidimaging.core.parallel import two_shared_mem as ptsm
