+++ mantidimaging/test/file_outputting_test_case.py
