+++ mantidimaging/test/filters_test/circular_mask_test.py
