+++ mantidimaging/test/filters_test/clip_values_test.py
