+++ mantidimaging/test/filters_test/crop_coords_test.py
