+++ mantidimaging/test/filters_test/gaussian_test.py
