+++ mantidimaging/test/filters_test/median_filter_test.py
