+++ mantidimaging/test/filters_test/minus_log_test.py
