+++ mantidimaging/test/filters_test/outliers_test.py
