+++ mantidimaging/test/filters_test/ring_removal_test.py
