+++ mantidimaging/test/filters_test/roi_normalisation_test.py
