+++ mantidimaging/test/filters_test/rotate_stack_test.py
