+++ mantidimaging/test/filters_test/stripe_removal_test.py
