+++ mantidimaging/test/test_helper.py
